refactor(plugins): address PR review on IR extension registration
- Remove the JacksonExtension IR extension and its tests; the runtime
  Wirespec Jackson module already handles refined/enum/record serialization.
  Drop its docs from plugins.md and the Polymorphic unions section of
  integration-jackson.mdx.
- Rename the plugin parameter extensionClasses -> extensions (Gradle task
  option + property, Maven @Parameter, and docs incl. Maven <extensions>/
  <extension> tags).
- Make the IrExtension classes final (drop 'open'):
  KotlinxSerializationExtension, SpringMappingAnnotationsExtension,
  SpringNativeHintsExtension.
- Rename extension tests dropping the redundant 'Ir'/language infix and
  move SpringJavaExtensionTest into the spring.extension package:
  KotlinxSerializationExtensionTest, SpringKotlinExtensionTest,
  SpringJavaExtensionTest.
