+++ src/integration/kotlinx-serialization/src/jvmTest/kotlin/community/flock/wirespec/integration/kotlinxserialization/extension/KotlinxSerializationExtensionTest.kt
@@ -17,7 +17,7 @@ import kotlin.test.Test
 import kotlin.test.assertContains
 import kotlin.test.assertFalse
 
-class KotlinxSerializationKotlinIrExtensionTest {
+class KotlinxSerializationExtensionTest {
 
     private fun parse(source: String): AST = object : ParseContext, NoLogger {
         override val spec = WirespecSpec
