+++ src/integration/spring/src/jvmMain/kotlin/community/flock/wirespec/integration/spring/extension/SpringNativeHintsExtension.kt
@@ -20,7 +20,7 @@ import community.flock.wirespec.ir.core.File as LanguageFile
  * Register alongside [SpringMappingAnnotationsExtension] on a Kotlin
  * or Java [community.flock.wirespec.ir.emit.IrEmitter].
  */
-open class SpringNativeHintsExtension(
+class SpringNativeHintsExtension(
     private val packageName: PackageName,
     private val language: FileExtension,
 ) : IrExtension {
