+++ src/site/docs/docs/integration/integration-jackson.mdx
@@ -179,40 +179,3 @@ public record Document(
     Boolean _final
 ) {}
 ```
-
-## Polymorphic unions
-
-A Wirespec `union` maps to a sealed interface with a record per member. Jackson needs to know how to pick
-the right concrete subtype when deserializing such a value. Rather than requiring a discriminator field in
-your contract, Wirespec ships an [IR extension](../plugins/plugins.md#ir-extensions),
-`JacksonExtension`, that annotates every generated union so the subtype is **deduced from the JSON
-shape**:
-
-```wirespec
-type UserAccount = UserAccountPassword | UserAccountToken
-type UserAccountPassword { username: String, password: String }
-type UserAccountToken { token: String }
-```
-
-is emitted with the polymorphism annotations injected before the sealed interface:
-
-```java
-@com.fasterxml.jackson.annotation.JsonTypeInfo(use = com.fasterxml.jackson.annotation.JsonTypeInfo.Id.DEDUCTION)
-@com.fasterxml.jackson.annotation.JsonSubTypes({
-    @com.fasterxml.jackson.annotation.JsonSubTypes.Type(value = UserAccountPassword.class),
-    @com.fasterxml.jackson.annotation.JsonSubTypes.Type(value = UserAccountToken.class)
-})
-public sealed interface UserAccount {}
-```
-
-Because `Id.DEDUCTION` is used, Jackson selects the member whose fields match the incoming JSON — no
-type discriminator is added to the wire format. The annotations are emitted fully qualified, so the
-generated code only needs `jackson-annotations` on its classpath. Records, refined types and enums need no
-extra annotations here — those cases are handled by the Wirespec Jackson module described above.
-
-:::note
-`JacksonExtension` works for both the Java and Kotlin targets and runs as part of the IR pipeline.
-Unlike the no-argument extensions, it takes the target language as a constructor argument, so it is wired
-into a custom `IrEmitter` rather than registered directly through the plugin's `extensionClasses` list.
-See [IR extensions](../plugins/plugins.md#ir-extensions) for the general mechanism.
-:::
