+++ src/site/docs/docs/plugins/plugins-maven.md
@@ -105,8 +105,8 @@ The compile mojo supports the following parameters:
 - **strict**: Whether to invoke strict mode during compilation. Default is 'true'.
 - **shared**: Whether to emit shared Wirespec code. Default is 'true'.
 - **emitterClass**: Specifies a custom emitter class to use for code generation.
-- **extensionClasses**: List of fully qualified `IrExtension` class names. The extensions are applied to the intermediate representation before code generation for every emitter that is an `IrEmitter`.
-- **ir**: Whether to emit through the intermediate representation. Required for `extensionClasses` to take effect on the built-in language targets. Default is 'false'.
+- **extensions**: List of fully qualified `IrExtension` class names. The extensions are applied to the intermediate representation before code generation for every emitter that is an `IrEmitter`.
+- **ir**: Whether to emit through the intermediate representation. Required for `extensions` to take effect on the built-in language targets. Default is 'false'.
 
 ### Running the Compile Goal
 
@@ -120,7 +120,7 @@ Or it will run automatically as part of the `generate-sources` phase during your
 
 ### Applying IR extensions
 
-[IR extensions](./plugins.md#ir-extensions) are registered with the `extensionClasses` parameter. Add the
+[IR extensions](./plugins.md#ir-extensions) are registered with the `extensions` parameter. Add the
 integration artifact that provides the extension as a plugin `<dependency>`, then list the fully qualified
 class name. Set `<ir>true</ir>` so the built-in language targets emit through the IR pipeline:
 
@@ -150,9 +150,9 @@ class name. Set `<ir>true</ir>` so the built-in language targets emit through th
                 </languages>
                 <ir>true</ir>
                 <shared>false</shared>
-                <extensionClasses>
-                    <extensionClass>community.flock.wirespec.integration.kotlinxserialization.extension.KotlinxSerializationExtension</extensionClass>
-                </extensionClasses>
+                <extensions>
+                    <extension>community.flock.wirespec.integration.kotlinxserialization.extension.KotlinxSerializationExtension</extension>
+                </extensions>
             </configuration>
         </execution>
     </executions>
